requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.29
alembic>=1.13.0
redis>=5.0.0
pyyaml>=6.0.1
structlog>=24.1.0
python-multipart>=0.0.9
pillow>=10.2.0
jsonschema>=4.21.0
httpx>=0.27.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
prometheus-client>=0.20.0
